fix: remove could have saved and saved from limit order success
