+++ src/components/MultiHopTrade/components/TradeSuccess/TradeSuccess.tsx
@@ -13,7 +13,7 @@ import {
   Text as CText,
   useDisclosure,
 } from '@chakra-ui/react'
-import { foxAssetId, fromAssetId, toAccountId } from '@shapeshiftoss/caip'
+import { fromAssetId, toAccountId } from '@shapeshiftoss/caip'
 import type { Asset } from '@shapeshiftoss/types'
 import { TransferType } from '@shapeshiftoss/unchained-client'
 import type { InterpolationOptions } from 'node-polyglot'
@@ -21,8 +21,6 @@ import { useCallback, useMemo } from 'react'
 import { useTranslate } from 'react-polyglot'
 
 import { TwirlyToggle } from '../TwirlyToggle'
-import { YouCouldHaveSaved } from './components/YouCouldHaveSaved'
-import { YouSaved } from './components/YouSaved'
 
 import { Amount } from '@/components/Amount/Amount'
 import { AnimatedCheck } from '@/components/AnimatedCheck'
@@ -32,19 +30,14 @@ import { Text } from '@/components/Text'
 import { useTxDetails, useTxDetailsQuery } from '@/hooks/useTxDetails/useTxDetails'
 import { bnOrZero } from '@/lib/bignumber/bignumber'
 import { fromBaseUnit } from '@/lib/math'
-import { selectRelatedAssetIdsInclusiveSorted } from '@/state/slices/related-assets-selectors'
 import { selectMarketDataByAssetIdUserCurrency } from '@/state/slices/selectors'
 import {
   selectActiveQuote,
   selectConfirmedTradeExecution,
-  selectFirstHop,
   selectIsActiveQuoteMultiHop,
-  selectLastHop,
-  selectTradeQuoteAffiliateFeeAfterDiscountUserCurrency,
-  selectTradeQuoteAffiliateFeeDiscountUserCurrency,
 } from '@/state/slices/tradeQuoteSlice/selectors'
 import { serializeTxIndex } from '@/state/slices/txHistorySlice/utils'
-import { useAppSelector, useSelectorWithArgs } from '@/state/store'
+import { useAppSelector } from '@/state/store'
 
 export type TradeSuccessProps = {
   handleBack: () => void
@@ -56,6 +49,7 @@ export type TradeSuccessProps = {
   buyAsset?: Asset
   sellAmountCryptoPrecision?: string
   quoteBuyAmountCryptoPrecision?: string
+  extraContent?: JSX.Element
 }
 
 export const TradeSuccess = ({
@@ -68,6 +62,7 @@ export const TradeSuccess = ({
   sellAsset,
   buyAsset,
   quoteBuyAmountCryptoPrecision,
+  extraContent,
 }: TradeSuccessProps) => {
   const translate = useTranslate()
   const tradeQuote = useAppSelector(selectActiveQuote)
@@ -77,19 +72,8 @@ export const TradeSuccess = ({
     defaultIsOpen: false,
   })
 
-  const firstHop = useAppSelector(selectFirstHop)
-  const lastHop = useAppSelector(selectLastHop)
   const tradeExecution = useAppSelector(selectConfirmedTradeExecution)
   const isMultiHop = useAppSelector(selectIsActiveQuoteMultiHop)
-
-  const feeSavingUserCurrency = useAppSelector(selectTradeQuoteAffiliateFeeDiscountUserCurrency)
-
-  const affiliateFeeUserCurrency = useAppSelector(
-    selectTradeQuoteAffiliateFeeAfterDiscountUserCurrency,
-  )
-
-  const hasFeeSaving = !bnOrZero(feeSavingUserCurrency).isZero()
-  const couldHaveReducedFee = !hasFeeSaving && !bnOrZero(affiliateFeeUserCurrency).isZero()
 
   // Get the actual received amount from the buy transaction *if* we can
   // i.e if this isn't a swap to a manual receive addy
@@ -146,19 +130,6 @@ export const TradeSuccess = ({
       .times(buyAssetMarketDataUserCurrency?.price ?? 0)
       .toString()
   }, [buyAssetMarketDataUserCurrency, maybeExtraDeltaCryptoPrecision, buyAsset])
-
-  const relatedAssetIdsFilter = useMemo(
-    () => ({
-      assetId: foxAssetId,
-      onlyConnectedChains: false,
-    }),
-    [],
-  )
-
-  const relatedAssetIds = useSelectorWithArgs(
-    selectRelatedAssetIdsInclusiveSorted,
-    relatedAssetIdsFilter,
-  )
 
   const AmountsLine = useCallback(() => {
     if (!(sellAsset && buyAsset)) return null
@@ -233,24 +204,6 @@ export const TradeSuccess = ({
     maybeExtraDeltaCryptoPrecision,
   ])
 
-  // NOTE: This is a temporary solution to enable the Fox discount summary only if the user did NOT
-
-  // trade FOX. If a user trades FOX, the discount calculations will have changed from the correct
-  // values because the amount of FOX held in the wallet will have changed.
-  // See https://github.com/shapeshift/web/issues/8028 for more details.
-  const enableFoxDiscountSummary = useMemo(() => {
-    const didTradeFox = relatedAssetIds.some(assetId => {
-      return (
-        firstHop?.buyAsset.assetId === assetId ||
-        firstHop?.sellAsset.assetId === assetId ||
-        lastHop?.buyAsset.assetId === assetId ||
-        lastHop?.sellAsset.assetId === assetId
-      )
-    })
-
-    return !didTradeFox
-  }, [firstHop, lastHop, relatedAssetIds])
-
   return (
     <>
       <CardBody pb={4} px={0}>
@@ -268,12 +221,7 @@ export const TradeSuccess = ({
           <Button mt={4} size='lg' width='full' onClick={handleBack} colorScheme='blue'>
             {translate(buttonTranslation)}
           </Button>
-          {enableFoxDiscountSummary && hasFeeSaving && feeSavingUserCurrency && (
-            <YouSaved feeSavingUserCurrency={feeSavingUserCurrency} />
-          )}
-          {couldHaveReducedFee && affiliateFeeUserCurrency && (
-            <YouCouldHaveSaved affiliateFeeUserCurrency={affiliateFeeUserCurrency} />
-          )}
+          {extraContent}
         </Stack>
       </CardBody>
       {summaryTranslation && children && (
